backend/requirements.txt
# AccessWorld Backend Dependencies
# FastAPI + server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# HuggingFace / ML
transformers==4.40.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
accelerate==0.29.3
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
soundfile==0.12.1
sentencepiece==0.2.0
sacremoses==0.1.1
datasets==2.19.0
huggingface-hub==0.22.2
openai-whisper==20231117

# Audio processing
librosa==0.10.2
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
